Introduce Logger::Formatters::Structured
