+++ lib/dry/logger/entry.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "time"
 require "dry/logger/constants"
 
 module Dry
